Move collateral/dtcc deposit plugins under integrations/deposits
The collateral PaymentsPlugin and DTCC deposit + plan-approval plugins now
live alongside the wallet/pull/ota deposits in integrations/deposits, wired
through a single registerDeposits aggregator (mirroring registerTokenStandards).
registry.ts now composes just two registrars: token standards and deposits.
Behavior-preserving — same plugins, same gating, same order.
