+++ src/integrations/deposits/collateral/index.ts
@@ -4,9 +4,9 @@ import {
   WalletResolver as CollateralWalletResolver,
 } from "@owneraio/finp2p-ethereum-collateral";
 import { FinP2PContract } from "@owneraio/finp2p-contracts";
-import { WalletResolver as CustodyWalletResolver } from "../../services/direct";
-import { IntegrationContext } from "../registry";
-import { pooledProvider, pooledSigner } from "../signer-pool";
+import { WalletResolver as CustodyWalletResolver } from "../../../services/direct";
+import { IntegrationContext } from "../../registry";
+import { pooledProvider, pooledSigner } from "../../signer-pool";
 
 /**
  * Registers the Ownera triparty collateral PaymentsPlugin when
